Completed the infrastructure to allow commands to be executed against NanoDB programmatically via the NanoDBServer class.  Once this was completed, it enables straightforward integration testing of the database server.  To this end, several additional classes were added under the test/edu/caltech/test/nanodb/sql directory to support easy construction of integration tests.
There were also a few changes necessary to support constructing tuple-literals from a set of values (to specify the expected results for a test), and comparing two tuples together (to see if expected and actual tuples actually matched); see the changes in src/edu/caltech/nanodb/expressions for details.

Finally, the students get a present - a brand new SortNode to use for ORDER BY expressions.