Updated plan nodes so that they would print "cost unknown" if the cost is unavailable.
Also added a workaround so that commands run through the NanoDBServer class will be able to read their table data properly.  There is a bug when executing joins via the NanoDBServer class that seems to do with the Storage Manager not caching valid schema information for the table.  Somehow it is becoming corrupted.  This workaround will allow people to work until the actual bug is tracked down.