Changes and updates necessary to support write-ahead logging and transaction processing.
The buffer manager is more sophisticated now, allowing pages to be pinned and unpinned.  Pages can also be released back to the storage manager, which causes them to become unpinned, and causes the write-ahead log to be updated if the page is dirty.  The buffer manager follows the write-ahead logging rule in writing pages out to disk; the WAL must be updated first, before data pages may be output.

Because the buffer manager and storage manager are much more intelligent now, NanoDB doesn't flush all tables to disk after every command.  Thus, it should actually behave reasonably efficiently now.

A new "CRASH" command was added to simulate simple crashing scenarios to test the write-ahead logger.  This will require a clean rebuild of the sources.