Fixed a number of bugs with the storage layer that were making write-ahead logging break.
The DBFileReader/Writer weren't properly mapping file-position into a page # and page offset, so this is fixed.

The StorageManager and DBPage now allow more fine-grained reporting of changes, and also allow unpinning of pages separate from reporting of writes.

The write-ahead log format has had a few missing parts filled in.  Also, the default value for the LogSequenceNumber recordSize value is 0, instead of -1.